app/consumer/requirements.txt
aiofiles==0.4.0
httptools==0.0.11
kafka-python==1.4.4
multidict==4.5.2
peewee==3.8.0
sanic==0.8.3
SQLAlchemy==1.2.15
ujson==1.35
uvloop==0.11.3
websockets==5.0.1
redis==2.10.5